+++ crates/runner-opentitan/src/board.rs
@@ -15,8 +15,7 @@
 use core::cell::RefCell;
 
 use critical_section::Mutex;
-use wasefire_board_api::{self as board, Event};
-use wasefire_error::Error;
+use wasefire_board_api::{self as board, AppletMemory, Event, Failure};
 use wasefire_scheduler as scheduler;
 
 mod applet;
@@ -78,12 +77,13 @@ impl board::Api for Board {
         }
     }
 
-    fn syscall(x1: u32, x2: u32, x3: u32, x4: u32) -> Option<Result<u32, Error>> {
+    fn vendor(mem: impl AppletMemory, x1: u32, x2: u32, x3: u32, x4: u32) -> Result<u32, Failure> {
+        #[cfg(not(feature = "test-vendor"))]
+        let _ = &mem;
         match (x1, x2, x3, x4) {
-            // The syscall_test example relies on this.
             #[cfg(feature = "test-vendor")]
-            (0, 0, 0, x) => Some(Error::decode(x as i32)),
-            _ => None,
+            (0, _, _, _) => syscall_test::process(mem, x2, x3, x4),
+            _ => Err(Failure::TRAP),
         }
     }