+++ crates/scheduler/src/call/crypto/ed25519.rs
@@ -19,14 +19,14 @@ use wasefire_board_api::Api as Board;
 #[cfg(feature = "board-api-crypto-ed25519")]
 use wasefire_board_api::crypto::ed25519::Api as _;
 #[cfg(feature = "board-api-crypto-ed25519")]
-use wasefire_board_api::{self as board, Support};
+use wasefire_board_api::{self as board, AppletMemory as _, AppletMemoryExt as _, Support};
 #[cfg(feature = "board-api-crypto-ed25519")]
 use wasefire_error::{Code, Error};
 
 #[cfg(feature = "board-api-crypto-ed25519")]
 use crate::Failure;
 #[cfg(feature = "board-api-crypto-ed25519")]
-use crate::applet::store::{Memory, MemoryApi};
+use crate::applet::store::Memory;
 use crate::{DispatchSchedulerCall, SchedulerCall};
 
 pub fn process<B: Board>(call: Api<DispatchSchedulerCall<B>>) {