+++ crates/scheduler/src/call/debug.rs
@@ -14,9 +14,8 @@
 
 use wasefire_applet_api::debug::{self as api, Api, Perf};
 use wasefire_board_api::debug::Api as _;
-use wasefire_board_api::{self as board, Api as Board};
+use wasefire_board_api::{self as board, Api as Board, AppletMemory as _, AppletMemoryExt as _};
 
-use crate::applet::store::MemoryApi;
 use crate::{DispatchSchedulerCall, SchedulerCall, Trap};
 
 pub fn process<B: Board>(call: Api<DispatchSchedulerCall<B>>) {