+++ crates/scheduler/src/call/store.rs
@@ -19,12 +19,12 @@ use wasefire_applet_api::store::Api;
 #[cfg(feature = "board-api-storage")]
 use wasefire_applet_api::store::{self as api};
 use wasefire_board_api::Api as Board;
+#[cfg(feature = "board-api-storage")]
+use wasefire_board_api::{AppletMemory as _, AppletMemoryExt as _};
 
 use crate::DispatchSchedulerCall;
 #[cfg(feature = "board-api-storage")]
 use crate::SchedulerCall;
-#[cfg(feature = "board-api-storage")]
-use crate::applet::store::MemoryApi;
 
 #[cfg(feature = "applet-api-store-fragment")]
 mod fragment;