+++ crates/scheduler/src/call/usb/serial.rs
@@ -19,11 +19,11 @@ use wasefire_applet_api::usb::serial::Api;
 use wasefire_board_api as board;
 use wasefire_board_api::Api as Board;
 #[cfg(feature = "board-api-usb-serial")]
+use wasefire_board_api::AppletMemory as _;
+#[cfg(feature = "board-api-usb-serial")]
 use wasefire_board_api::usb::serial::{Api as _, Event};
 
 use crate::DispatchSchedulerCall;
-#[cfg(feature = "board-api-usb-serial")]
-use crate::applet::store::MemoryApi;
 #[cfg(feature = "board-api-usb-serial")]
 use crate::event::{Handler, usb::serial::Key};
 #[cfg(feature = "board-api-usb-serial")]