+++ examples/rust/syscall_test/src/lib.rs
@@ -15,10 +15,15 @@
 //! Tests that the board-specific syscall API is working properly.
 
 #![no_std]
+
+use alloc::vec;
 wasefire::applet!();
 
 fn main() {
     test_error();
+    test_read();
+    test_write();
+    test_alloc();
     scheduling::exit();
 }
 
@@ -33,5 +38,36 @@ fn test_error() {
     ] {
         debug!("- {x:08x} -> {r:?}");
         assert_eq!(unsafe { syscall(0, 0, 0, x) }, r);
+    }
+}
+
+fn test_read() {
+    debug!("test_read(): Check that platform can read memory.");
+    for len in [0, 1, 7, 8] {
+        let data = rng::bytes(len).unwrap();
+        debug!("- {data:02x?}");
+        let expected = data.iter().map(|&x| x as usize).sum();
+        let actual = unsafe { syscall(0, 1, len, data.as_ptr().addr()) }.unwrap();
+        assert_eq!(actual, expected);
+    }
+}
+
+fn test_write() {
+    debug!("test_write(): Check that platform can write memory.");
+    for len in [0, 1, 7, 8] {
+        let mut data = vec![0u8; len];
+        assert_eq!(unsafe { syscall(0, 2, len, data.as_mut_ptr().addr()) }.unwrap(), len);
+        debug!("- {data:02x?}");
+        assert!(data.iter().zip((0 .. len).rev()).all(|(&x, y)| x == y as u8));
+    }
+}
+
+fn test_alloc() {
+    debug!("test_alloc(): Check that platform can allocated memory.");
+    for len in [1, 7, 8, 9] {
+        let ptr = unsafe { syscall(0, 3, len, 0) }.unwrap();
+        let data = unsafe { core::slice::from_raw_parts(ptr as *const u8, len) };
+        debug!("- {data:02x?}");
+        assert!(data.iter().all(|&x| x == len as u8));
     }
 }