+++ crates/scheduler/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Minor
 
+- Use `Trap`, `Failure`, and `AppletMemory{,Ext}` from `wasefire-board-api`
 - Support `fingerprint::{matcher,sensor}` modules
 - Support `PlatformClearStore` protocol call
 - Support `crypto::{ecdsa,ecdh,ed25519}` modules